requirements.txt
chainlit==1.0.0
fastapi==0.100.1
uvicorn==0.23.2
transformers==4.35.2
datasets==2.20.0
accelerate==0.25.0
peft==0.6.2
pytesseract==0.3.10
pillow==10.1.0
opencv-python==4.8.1.78
rapidfuzz==3.5.2
pydantic==2.7.3
tinydb==4.8.0 
pytest==7.4.3
torch==1.13.1
label-studio==1.13.1 
albumentations==1.3.1 